[m4/restructure] reorganized files
